[gh-#273] move activity service
